refactor: root page를 서버컴포넌트로 변경 (추후에 prefetch를 고려하여 변경함)
